When the port opens/closes
Changed when the port opens and closes.
Also changed the return dictionary types when the serial gives a
response.

+++ BrewduinoCatalogLib/IArduinoSerial.cs
@@ -9,6 +9,6 @@ namespace BrewduinoCatalogLib
     {
 
         void SendCommand(ArduinoCommands.CommandTypes cmd, string text);
-        Dictionary<string, int> SendCommandWithResponse(ArduinoCommands.CommandTypes cmd, string text);
+        Dictionary<string, decimal> SendCommandWithResponse(ArduinoCommands.CommandTypes cmd, string text);
     }
 }
